tests: refactor core multisite test helpers into constants and utils
