+++ rdmo/management/tests/test_viewset_import_multisite.py
@@ -2,8 +2,8 @@ import pytest
 
 from django.urls import reverse
 
-from rdmo.core.tests import get_obj_perms_status_code
-from rdmo.core.tests import multisite_users as users
+from rdmo.core.tests.constants import multisite_users as users
+from rdmo.core.tests.utils import get_obj_perms_status_code
 from rdmo.questions.models import Catalog, Page, Question, QuestionSet, Section
 
 status_map = {
