+++ rdmo/options/tests/test_viewset_optionsets_multisite.py
@@ -4,9 +4,10 @@ import pytest
 
 from django.urls import reverse
 
-from ...core.tests import get_obj_perms_status_code
-from ...core.tests import multisite_status_map as status_map
-from ...core.tests import multisite_users as users
+from rdmo.core.tests.constants import multisite_status_map as status_map
+from rdmo.core.tests.constants import multisite_users as users
+from rdmo.core.tests.utils import get_obj_perms_status_code
+
 from ..models import OptionSet
 from .test_viewset_optionsets import urlnames
